+++ cloudFoundry/kafana-spring-mongodb/src/main/java/be/kafana/foursquare/down/VenueGrabber.java
@@ -9,13 +9,24 @@ import be.kafana.foursquare.down.data.Venue;
 
 public interface VenueGrabber {
 
+  /**
+   * 
+   * @param startLatitiude
+   * @param startLongitude
+   * @param endLatitude
+   * @param endLongitude
+   * @param rate is the rate in witch longitude and latitude will be increased
+   * @param maxItems maximal number of Venue's that will be gathered
+   * @return
+   */
   public abstract Set<Venue> grab(double startLatitiude,
       double startLongitude,
       double endLatitude,
       double endLongitude,
-      int rate,
+      double rate,
       long maxItems);
-
+   
+  
   public abstract OAuthData getData();
 
   public abstract void setData(OAuthData data);
